requisitos 11 e 12

+++ back-end/src/database/models/Product.js
@@ -4,7 +4,7 @@ module.exports = (sequelize, DataTypes) => {
     {
       id: { primaryKey: true, type: DataTypes.INTEGER, autoIncrement: true },
       name: DataTypes.STRING,
-      price: DataTypes.FLOAT,
+      price: DataTypes.DECIMAL(10, 2),
       urlImage: DataTypes.STRING,
     },
     {
